requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
django-environ